[REF] core: break odoo.orm.fields into multiple files
Fields contained a lot of different classes and was the biggest file in
the ORM. Splitting it also allows to define Command, Field (meta field),
scalar fields and relational fields in different files. This solves some
cyclic dependencies with models: now we can import scalar fields, then
models and, at the end, the relational fields.
