+++ odoo/models/__init__.py
@@ -5,7 +5,7 @@
 # TODO we should only expose *Model objects here, maybe check_comp*
 from odoo.tools import _
 
-from odoo.orm.fields import Command
+from odoo.orm.commands import Command
 from odoo.orm.identifiers import NewId
 from odoo.orm.models import (
     GC_UNLINK_LIMIT,
